requirements.txt
requests
google-cloud-storage
Flask
selenium
webdriver-manager
gunicorn
